requirements.txt
fastapi
uvicorn
python-multipart
GitPython
pydantic-settings
pika
prometheus-client